Identify Deme Inbreeding Spatial Coefficients in Continuous Space

The purpose of this statistic is to identify an inbreeding coefficient, or degree of relatedness, for a given location in space. We assume that locations in spaces can be represented as "demes," such that multiple individuals live in the same deme (i.e. samples are sourced from the same location). The expected pairwise relationship between two individuals, or samples, is dependent on the each sample's deme's inbreeding coefficient and the geographic distance between the demes. The program assumes a symmetric distance matrix.

deme_inbreeding_spcoef(
  discdat,
  start_params = c(),
  f_learningrate = 1e-05,
  m_learningrate = 1e-10,
  momentum = 0.9,
  steps = 1000,
  report_progress = TRUE,
  return_verbose = FALSE
)

Arguments

discdat

dataframe; The genetic-geographic data by deme (K)

start_params

named double vector; vector of start parameters.

f_learningrate

double; alpha parameter for how much each "step" is weighted in the gradient descent for inbreeding coefficients

m_learningrate

double; alpha parameter for how much each "step" is weighted in the gradient descent for the migration parameter

momentum

double; gamma parameter for momentum for adaptive learning rate

steps

integer; the number of "steps" as we move down the gradient

report_progress

boolean; whether or not a progress bar should be shown as you iterate through steps

return_verbose

boolean; whether the inbreeding coefficients and migration rate should be returned for every iteration or only for the final iteration. User will typically not want to store every iteration, which can be memory intensive

Details

The gen.geo.dist dataframe must be named with the following columns: "smpl1"; "smpl2"; "deme1"; "deme2"; "gendist"; "geodist"; which corresponds to: Sample 1 Name; Sample 2 Name; Sample 1 Location; Sample 2 Location; Pairwise Genetic Distance; Pairwise Geographpic Distance. Note, the order of the columns do not matter but the names of the columns must match.

The start_params vector names must match the cluster names (i.e. clusters must be have a name that we can match on for the starting relatedness paramerts). In addition, you must provide a start parameter for "m".

Note: We have implemented coding decisions to not allow the "f" inbreeding coefficients to be negative by using a logit transformation internally in the code. Similarly, the "m" global migration rate cannot be negative and is bounded to be positive by through taking the absolute value in every iteration.