Find Optimal Gradient Parameters

Find Optimal Gradient Parameters

find_grad_params(
  discdat,
  initstart_params = c(),
  initf_learningrate = 1e-05,
  initm_learningrate = 1e-10,
  momentum = 0.9,
  discsteps = 1000,
  initTemp = 1,
  annealsteps = 1e+05,
  SLratio = 0.5,
  FMSratio = 0.5,
  FMLratio = 0.5,
  demeSwitchSize = 1,
  fstartmin = 0.1,
  fstartmax = 0.9,
  fstartsteps = 100,
  mstartmin = 1e-15,
  mstartmax = 0.001,
  mstartsteps = 100,
  flearnmin = 1e-15,
  flearnmax = 0.01,
  flearnsteps = 100,
  mlearnmin = 1e-15,
  mlearnmax = 0.01,
  mlearnsteps = 100
)

Arguments

discdat

dataframe; The genetic-geographic data by deme (K)

initstart_params

named double vector; vector of start parameters.

initf_learningrate

double; alpha parameter for how much each "step" is weighted in the gradient descent for inbreeding coefficients

initm_learningrate

double; alpha parameter for how much each "step" is weighted in the gradient descent for the migration parameter

momentum

double; gamma parameter for momentum for adaptive learning rate

discsteps

integer; the number of "steps" in the main DISCent algorithm" deme_inbreeding_spcoef

initTemp

double; initial temperature for simulated annealer

annealsteps

integer; the number of "steps" for the simulated annealer to consider for identifying optimal start and learning parameters

SLratio

double; probability of searching for start parameters versus the learning rate

FMSratio

double; probability of searching for F start parameters versus the M start parameters

FMLratio

double; probability of searching for F learning rate versus the M learning rate

demeSwitchSize

integer; the number of demes to "move" in simulated annealing proposal step

fstartmin

double; the number of values of F inbreeding coefficient to explore in the search grid

fstartmax

double; the maximum value for F inbreeding coefficient in the search grid

fstartsteps

integer; the number of values between fstartmin and fstartmax to evaluate

mstartmin

double; the number of values of M global migration rate to explore in the search grid

mstartmax

double; the maximum value for M global migration rate in the search grid

mstartsteps

integer; the number of values between mstartmin and mstartmax to evaluate

flearnmin

double; the number of values of F learning rate to explore in the search grid

flearnmax

double; the maximum value for F learning rate in the search grid

flearnsteps

integer; the number of values between flearnmin and flearnmax to evaluate

mlearnmin

double; the number of values of M learning rate to explore in the search grid

mlearnmax

double; the maximum value for M learning rate in the search grid

mlearnsteps

integer; the number of values between mlearnmin and mlearnmax to evaluate

Details

Using simulated annealing, identify the best start parameters for inbreeding F values and M global migration parameter as well as the best learning rates. The simulated annealer assumes a discrete search grid and temperate decreases according to the rate: t = fractiter The cost in the simulated annealer is given by the final cost after n iterations of the main DISCent algorithm" deme_inbreeding_spcoef. The search grid is defined by the user specified min, max, and steps of the aforementioned parameters. The various ratios of searching for the "X" parameters versus the "Y" parameters helps to prioritize exploring one set of parameters versus the other. The default value of 0.5 leads to equal probability of exploration.