DISCent Forward Simulator

Uses the DISCent equation, \(r_ij = (f_i + f_j/2)e^(-d_{ij}/m)\) to simulate data expected genetic distances between sample pairs within demes. We assume a beta-binomial model to create realistic "noise" amongst the pairs, where \(Y|p ~ binom(n,p)\) such that \(p ~ beta(\mu * \phi, (1 - \mu) * \phi)\), where \(\mu\) is average relatedness (defined by \(\rij\)) and \(\phi\) is a concentration parameter of that relatedness. In our instance the beta-binomial returns counts (ie number of related segments), we fix the segment denominator at 100. Overdispersion can be adjusted to produce more or less correlation of relatedness

run_forward_disc(
  true_params,
  geodist_matrix,
  samples_per_deme,
  overdispersion = 200
)

Arguments

true_params

named vector; True F_i and M values to simulate. At least one element must be named "m".

geodist_matrix

named matrix of geodistances; must be a nxn geodistance matrix, with n corresponding to number of demes and deme names matching the true_param names

samples_per_deme

integer vector; number of pairs within each deme

overdispersion

numeric; overdispersion parameter in the beta-binomial model; Default: 200

Value

A dataframe of class "DISCsim" containing columns:

Final_Fis: Final inbreeding coefficient estimates

Final_m: Final migration rate estimate

deme_key: Mapping of deme names to indices

cost: Final cost function value(s)